+++ lib/valuables/time_duration_response.rb
@@ -8,13 +8,16 @@ module Valuables
 
     def name
       hash = parse_time_duration(@object.response)
+      h = (hash[:hours] == 1 ? 'hour' : 'hours')
+      m = (hash[:minutes] == 1 ? 'minute' : 'minutes')
+      s = (hash[:seconds] == 1 ? 'second' : 'seconds')
       case @object.variable.time_duration_format
       when 'mm:ss'
-        "#{hash[:minutes]}m #{hash[:seconds]}s"
+        "#{hash[:minutes]} #{m} #{hash[:seconds]} #{s}"
       when 'hh:mm'
-        "#{hash[:hours]}h #{hash[:minutes]}m"
-      else # 'hh:mm:ss'
-        "#{hash[:hours]}h #{hash[:minutes]}m #{hash[:seconds]}s"
+        "#{hash[:hours]} #{h} #{hash[:minutes]} #{m}"
+      else
+        "#{hash[:hours]} #{h} #{hash[:minutes]} #{m} #{hash[:seconds]} #{s}"
       end
     rescue
       @object.response